Fix concurrent mode bug
If you useFirestoreDoc to subscribe to a document, and then to
another, it would stay subscribed to the first. The test I added doesn't
catch this bug, so it's not really a regression test, but it might catch
similar issues. Maybe the renderer used in the testing framework works
differently.